refactored for export ClientUtils
