+++ lib/cjs/Launcher/SmartLaunchHandler.js
@@ -71,11 +71,9 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getEndpointsForEmr = exports.instanceOfEmr = exports.EMR = void 0;
+exports.instanceOfEmr = exports.EMR = void 0;
 var FHIR = __importStar(require("fhirclient"));
-var CernerClient_1 = __importDefault(require("../Client/CernerClient"));
 var ClientFactory_1 = require("../Client/ClientFactory");
-var EpicClient_1 = __importDefault(require("../Client/EpicClient"));
 var Config_1 = require("./Config");
 var scopes_json_1 = __importDefault(require("./scopes.json"));
 var EMR;
@@ -94,25 +92,6 @@ function instanceOfEmr(object) {
     return Object.values(EMR).includes(object);
 }
 exports.instanceOfEmr = instanceOfEmr;
-/**
-* The function `getEndpointsForEmr` returns the endpoints for a given EMR type, such as Epic, Cerner, or SMART.
-* @param {EMR} emrType - The `emrType` parameter is of type `EMR`, which is an enumeration representing different types of Electronic Medical Record (EMR)
-* systems. The possible values for `emrType` are `EMR.EPIC`, `EMR.CERNER`, `EMR.SMART`,
-* @returns an object of type EMR_ENDPOINTS.
-*/
-function getEndpointsForEmr(emrType) {
-    switch (emrType) {
-        case EMR.EPIC:
-            return EpicClient_1.default.getEndpoints();
-        case EMR.CERNER:
-            return CernerClient_1.default.getEndpoints();
-        case EMR.SMART:
-        case EMR.NONE:
-        default:
-            throw new Error("Endpoints not found for EMR type: ".concat(emrType));
-    }
-}
-exports.getEndpointsForEmr = getEndpointsForEmr;
 /**
  * Represents the SmartLaunchHandler class.
  */
