+++ lib/cjs/index.js
@@ -26,14 +26,15 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getEndpointsForEmr = exports.LAUNCH = exports.EMR = exports.BaseClient = exports.SmartLaunchHandler = exports.ClientFactory = void 0;
+exports.ClientUtils = exports.SmartLaunchHandler = exports.LAUNCH = exports.EMR = exports.ClientFactory = exports.BaseClient = void 0;
+var BaseClient_1 = __importDefault(require("./Client/BaseClient"));
+exports.BaseClient = BaseClient_1.default;
 var ClientFactory_1 = __importStar(require("./Client/ClientFactory"));
 exports.ClientFactory = ClientFactory_1.default;
 Object.defineProperty(exports, "LAUNCH", { enumerable: true, get: function () { return ClientFactory_1.LAUNCH; } });
+var ClientUtils = __importStar(require("./Client/utils"));
+exports.ClientUtils = ClientUtils;
 var SmartLaunchHandler_1 = __importStar(require("./Launcher/SmartLaunchHandler"));
 exports.SmartLaunchHandler = SmartLaunchHandler_1.default;
 Object.defineProperty(exports, "EMR", { enumerable: true, get: function () { return SmartLaunchHandler_1.EMR; } });
-Object.defineProperty(exports, "getEndpointsForEmr", { enumerable: true, get: function () { return SmartLaunchHandler_1.getEndpointsForEmr; } });
-var BaseClient_1 = __importDefault(require("./Client/BaseClient"));
-exports.BaseClient = BaseClient_1.default;
 //# sourceMappingURL=index.js.map
