+++ lib/esm/Launcher/SmartLaunchHandler.js
@@ -8,9 +8,7 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 import * as FHIR from "fhirclient";
-import CernerClient from "../Client/CernerClient";
 import { LAUNCH } from "../Client/ClientFactory";
-import EpicClient from "../Client/EpicClient";
 import { cerner } from "./Config";
 import scopes from "./scopes.json";
 export var EMR;
@@ -27,24 +25,6 @@ export var EMR;
  */
 export function instanceOfEmr(object) {
     return Object.values(EMR).includes(object);
-}
-/**
-* The function `getEndpointsForEmr` returns the endpoints for a given EMR type, such as Epic, Cerner, or SMART.
-* @param {EMR} emrType - The `emrType` parameter is of type `EMR`, which is an enumeration representing different types of Electronic Medical Record (EMR)
-* systems. The possible values for `emrType` are `EMR.EPIC`, `EMR.CERNER`, `EMR.SMART`,
-* @returns an object of type EMR_ENDPOINTS.
-*/
-export function getEndpointsForEmr(emrType) {
-    switch (emrType) {
-        case EMR.EPIC:
-            return EpicClient.getEndpoints();
-        case EMR.CERNER:
-            return CernerClient.getEndpoints();
-        case EMR.SMART:
-        case EMR.NONE:
-        default:
-            throw new Error(`Endpoints not found for EMR type: ${emrType}`);
-    }
 }
 /**
  * Represents the SmartLaunchHandler class.
