+++ src/Launcher/SmartLaunchHandler.ts
@@ -1,8 +1,5 @@
 import * as FHIR from "fhirclient";
-import { EMR_ENDPOINTS } from "../Client/BaseClient";
-import CernerClient from "../Client/CernerClient";
 import { LAUNCH } from "../Client/ClientFactory";
-import EpicClient from "../Client/EpicClient";
 import { cerner } from "./Config";
 import scopes from "./scopes.json";
 
@@ -20,25 +17,6 @@ export enum EMR {
  */
 export function instanceOfEmr(object: unknown): object is EMR {
   return Object.values(EMR).includes((object as string) as EMR)
-}
-
-/**
-* The function `getEndpointsForEmr` returns the endpoints for a given EMR type, such as Epic, Cerner, or SMART.
-* @param {EMR} emrType - The `emrType` parameter is of type `EMR`, which is an enumeration representing different types of Electronic Medical Record (EMR)
-* systems. The possible values for `emrType` are `EMR.EPIC`, `EMR.CERNER`, `EMR.SMART`,
-* @returns an object of type EMR_ENDPOINTS.
-*/
-export function getEndpointsForEmr(emrType: EMR): EMR_ENDPOINTS {
-  switch (emrType) {
-    case EMR.EPIC:
-      return EpicClient.getEndpoints()
-    case EMR.CERNER:
-      return CernerClient.getEndpoints()
-    case EMR.SMART:
-    case EMR.NONE:
-    default:
-      throw new Error(`Endpoints not found for EMR type: ${emrType}`)
-  }
 }
 
 /**
@@ -198,7 +176,7 @@ export default class SmartLaunchHandler {
           ? redirectPath
           : '/' + redirectPath
         : ''
-      );
+    );
     const urlParams = new URLSearchParams(queryString);
     const iss = urlParams.get("iss") ?? undefined;
     if (!iss)
