+++ src/index.ts
@@ -1,4 +1,6 @@
-import ClientFactory, {LAUNCH} from "./Client/ClientFactory";
-import SmartLaunchHandler, { EMR, getEndpointsForEmr } from "./Launcher/SmartLaunchHandler";
 import BaseClient from "./Client/BaseClient";
-export { ClientFactory, SmartLaunchHandler, BaseClient, EMR, LAUNCH, getEndpointsForEmr };
+import ClientFactory, { LAUNCH } from "./Client/ClientFactory";
+import * as ClientUtils from "./Client/utils";
+import SmartLaunchHandler, { EMR } from "./Launcher/SmartLaunchHandler";
+export { BaseClient, ClientFactory, EMR, LAUNCH, SmartLaunchHandler, ClientUtils };
+
